Move blocking I/O in app list rendering to background thread
Items in the app list views need to load the icon URL and version
information during rendering because it's not available in the cursor
loaded by the CursorLoader. They do this by requesting it from the
corresponding App object, which issues a database query (per item)
on the UI thread, blocking the UI.

With this change, these queries are done on a background thread.
The UI thread then updates the icon and version fields when the
data becomes available.